Simplify advisor code
