+++ rafiki/advisor/service.py
@@ -4,9 +4,8 @@ import os
 import uuid
 import traceback
 import pprint
-import numpy as np
 
-from .advisor import make_advisor
+from .advisor import Advisor
 
 logger = logging.getLogger(__name__)
 
@@ -25,12 +24,13 @@ class AdvisorService(object):
             advisor = self._get_advisor(advisor_id)
             
         if advisor is None:
-            advisor_inst = make_advisor(knob_config)
-            advisor = self._create_advisor(advisor_inst, knob_config, advisor_id)
+            advisor = Advisor(knob_config)
+            advisor_id = str(uuid.uuid4()) if advisor_id is None else advisor_id
+            self._advisors[advisor_id] = advisor
             is_created = True
 
         return {
-            'id': advisor.id,
+            'id': advisor_id,
             'is_created': is_created # Whether a new advisor has been created
         }
 
@@ -40,7 +40,7 @@ class AdvisorService(object):
         advisor = self._get_advisor(advisor_id)
 
         if advisor is not None:
-            self._delete_advisor(advisor)
+            del self._advisors[advisor_id]
             is_deleted = True
 
         return {
@@ -51,7 +51,7 @@ class AdvisorService(object):
 
     def generate_proposal(self, advisor_id):
         advisor = self._get_advisor(advisor_id)
-        knobs = self._generate_proposal(advisor)
+        knobs = advisor.propose()
 
         return {
             'knobs': knobs
@@ -65,18 +65,12 @@ class AdvisorService(object):
         if advisor is None:
             raise InvalidAdvisorException()
 
-        advisor_inst = advisor.advisor_inst
-        advisor_inst.feedback(knobs, score)
-        knobs = self._generate_proposal(advisor)
+        advisor.feedback(knobs, score)
+        knobs = advisor.propose()
 
         return {
             'knobs': knobs
         }
-
-    def _create_advisor(self, advisor_inst, knob_config, advisor_id=None):
-        advisor = Advisor(advisor_inst, knob_config, advisor_id)
-        self._advisors[advisor.id] = advisor
-        return advisor
 
     def _get_advisor(self, advisor_id):
         if advisor_id not in self._advisors:
@@ -84,39 +78,3 @@ class AdvisorService(object):
 
         advisor = self._advisors[advisor_id]
         return advisor
-
-    def _update_advisor(self, advisor, advisor_inst):
-        advisor.advisor_inst = advisor_inst
-        return advisor
-
-    def _delete_advisor(self, advisor):
-        del self._advisors[advisor.id]
-
-    def _generate_proposal(self, advisor):
-        knobs = advisor.advisor_inst.propose()
-
-        # Simplify knobs to use JSON serializable values
-        knobs = {
-            name: self._simplify_value(value)
-                for name, value
-                in knobs.items()
-        }
-
-        return knobs
-
-    def _simplify_value(self, value):
-        # TODO: Support int64 & other non-serializable data formats
-        if isinstance(value, np.int64):
-            return int(value)
-
-        return value
-
-class Advisor(object):
-    def __init__(self, advisor_inst, knob_config, advisor_id=None):
-        if advisor_id is not None:
-            self.id = advisor_id
-        else:
-            self.id = str(uuid.uuid4())
-        
-        self.advisor_inst = advisor_inst
-        self.knob_config = knob_config
