+++ rafiki/model/requirements.txt
@@ -2,3 +2,4 @@ numpy==1.14.5
 Pillow==5.3.0
 requests==2.20.0
 tqdm==4.28.1
+baytune==0.2.1
